# backend/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Projectiva API
  version: 1.0.0
  description: API documentation for the Projectiva project management application.
servers:
  - url: http://localhost:5002/api
    description: Local development server
paths:
  /projects:
    # Add a new GET method to retrieve all projects
    get:
      summary: Get all projects
      tags:
        - Projects
      responses:
        '200':
          description: Successful response - returns an array of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
    # Add a new POST method to create a new project
    post:
      summary: Create a new project
      tags:
        - Projects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
      responses:
        '201':
          description: Project created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Bad request (e.g., invalid input data)
...
  /tasks:
    # Add a new GET method to retrieve all tasks
    get:
      summary: Get all tasks
      tags:
        - Tasks
      responses:
        '200':
          description: Successful response - returns an array of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    # Add a new POST method to create a new task
    post:
      summary: Create a new task
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request (e.g., invalid input data)
  # Add a new path /tasks/{id}:
...
  /subtasks:
    # Add a new POST method to create a new subtask
    post:
      summary: Create a new subtask
      tags:
        - Subtasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Subtask'
      responses:
        '201':
          description: Subtask created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subtask'
        '400':
          description: Bad request (e.g., invalid input data)
  # Add a new path /subtasks/{id}:
...
  /clients:
    # Add a new GET method to retrieve all clients
    get:
      summary: Get all clients
      tags:
        - Clients
      responses:
        '200':
          description: Successful response - returns an array of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Client'
    # Add a new POST method to create a new client
    post:
      summary: Create a new client
      tags:
        - Clients
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Client'
      responses:
        '201':
          description: Client created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '400':
          description: Bad request (e.g., invalid input data)
  # Add a new path
...
  /technicians:
    # Add a new GET method to retrieve all technicians
    get:
      summary: Get all technicians
      tags:
        - Technicians
      responses:
        '200':
          description: Successful response - returns an array of technicians
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Technician'
    # Add a new POST method to create a new technician
    post:
      summary: Create a new technician
      tags:
        - Technicians
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Technician'
      responses:
        '201':
          description: Technician created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Technician'
        '400':
          description: Bad request (e.g., invalid input data)
  # Add a new path /technicians/{id}:
...
  /equipment:
    # Add a new GET method to retrieve all equipment
    get:
      summary: Get all equipment
      tags:
        - Equipment
      responses:
        '200':
          description: Successful response - returns an array of equipment
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Equipment'
    # Add a new POST method to create a new equipment
    post:
      summary: Create new equipment
      tags:
        - Equipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Equipment'
      responses:
        '201':
          description: Equipment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipment'
        '400':
          description: Bad request (e.g., invalid input data)
  # Add a new path /equipment/{id}:
...
  /tools:
    # Add a new GET method to retrieve all tools
    get:
      summary: Get all tools
      tags:
        - Tools
      responses:
        '200':
          description: Successful response - returns an array of tools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tool'
    # Add a new POST method to create a new tool
    post:
      summary: Create new tool
      tags:
        - Tools
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tool'
      responses:
        '201':
          description: Tool created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tool'
        '400':
          description: Bad request (e.g., invalid input data)
  # Add a new path /tools/{id}:
...
  /notifications:
    # Add a new GET method to retrieve all notifications
    get:
      summary: Get all notifications
      tags:
        - Notifications
      responses:
        '200':
          description: Successful response - returns an array of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
    # Add a new POST method to create a new notification
    post:
      summary: Create a new notification
      tags:
        - Notifications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Notification'
      responses:
        '201':
          description: Notification created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        '400':
          description: Bad request (e.g., invalid input data)
  # Add a new path /notifications/{id}:
...
  /landing-content:
    get:
      summary: Get all landing page content
      tags:
        - Landing Content
      responses:
        '200':
          description: Successful response - returns an array of landing content items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LandingContent'
    post:
      summary: Create new landing page content
      tags:
        - Landing Content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LandingContent'
      responses:
        '201':
          description: Landing content created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LandingContent'
        '400':
          description: Bad request (e.g., invalid input data)
...
  /webhooks:
    post:
      summary: Handle Clerk webhooks
      tags:
        - Webhooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  description: The type of webhook event (e.g., 'user.created', 'user.updated').
                  example: "user.created"
                data:
                  type: object
                  description: The data payload of the webhook event.
                  # You might want to define a more specific schema for the data payload based on the expected events.
      responses:
        '200':
          description: Webhook received and processed successfully
        '400':
          description: Webhook verification failed
        '500':
          description: Internal server error
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
          description: The project ID.
        name:
          type: string
          description: The name of the project.
        type:
          type: string
          description: The type of project.
        department:
          type: string
          description: The department the project belongs to.
        client_id:
          type: integer
          description: The ID of the client associated with the project.
        location:
          type: string
          description: The location of the project.
        start_date:
          type: string
          format: date
          description: The start date of the project.
        end_date:
          type: string
          format: date
          description: The end date of the project.
        status:
          type: string
          description: The status of the project (e.g., Pending, In Progress, Completed).
        created_at:
          type: string
          format: date-time
          description: The timestamp when the project was created.
        updated_at:
          type: string
          format: date-time
          description: The timestamp when the project was last updated.
      example:
        id: 1
        name: Example Project
        type: Web Development
        department: IT
        client_id: 2
        location: Remote
        start_date: 2024-01-15
        end_date: 2024-05-30
        status: Pending
        created_at: 2024-01-10T08:00:00Z
        updated_at: 2024-01-10T08:00:00Z
    # Define the Task schema
    Task:
      type: object
      properties:
        id:
          type: integer
          description: The task ID.
          example: 1
        project_id:
          type: integer
          description: The ID of the project this task belongs to.
          example: 2
        name:
          type: string
          description: The name of the task.
          example: "Implement user authentication"
        assigned_to:
          type: integer
          description: The ID of the user or technician assigned to the task.
          example: 3
        due_date:
          type: string
          format: date
          description: The due date of the task.
          example: "2024-04-15"
        status:
          type: string
          description: The status of the task (e.g., To Do, In Progress, Completed).
          example: "In Progress"
        priority:
          type: string
          description: The priority of the task (e.g., Low, Medium, High).
          example: "High"
        description:
          type: string
          description: A detailed description of the task.
          example: "Implement user authentication using Clerk and OAuth 2.0."
        attachments:
          type: array
          description: An array of attachments associated with the task.
          items:
            type: object
            properties:
              name:
                type: string
                description: The name of the attachment.
                example: "design_document.pdf"
              url:
                type: string
                description: The URL where the attachment can be accessed.
                example: "https://example.com/attachments/design_document.pdf"
        created_at:
          type: string
          format: date-time
          description: The timestamp when the task was created.
          example: "2024-03-15T09:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: The timestamp when the task was last updated.
          example: "2024-03-15T14:45:00Z"
      required:
        - name
        - project_id
    # Define the Subtask schema
    Subtask:
        type: object
        properties:
          id:
            type: integer
            description: The subtask ID.
            example: 1
          task_id:
            type: integer
            description: The ID of the task this subtask belongs to.
            example: 2
          name:
            type: string
            description: The name of the subtask.
            example: "Set up database connection"
          is_completed:
            type: boolean
            description: Whether the subtask is completed or not.
            example: false
          created_at:
            type: string
            format: date-time
            description: The timestamp when the subtask was created.
            example: "2024-03-15T10:00:00Z"
          updated_at:
            type: string
            format: date-time
            description: The timestamp when the subtask was last updated.
            example: "2024-03-15T11:00:00Z"
        required:
          - name
          - task_id
    # Define the Client schema
    Client:
      type: object
      properties:
        id:
          type: integer
          description: The client ID.
          example: 1
        name:
          type: string
          description: The name of the client.
          example: "Acme Corporation"
        industry:
          type: string
          description: The industry the client belongs to.
          example: "Technology"
        contact_info:
          type: object
          description: Contact information for the client.
          properties:
            email:
              type: string
              format: email
              example: "info@acme.com"
            phone:
              type: string
              example: "555-123-4567"
        location:
          type: object
          description: Location information for the client.
          properties:
            address:
              type: string
              example: "123 Main St"
            city:
              type: string
              example: "Anytown"
            state:
              type: string
              example: "CA"
            zip:
              type: string
              example: "90210"
        notes:
          type: string
          description: Any notes about the client.
          example: "Important client for the company."
        created_at:
          type: string
          format: date-time
          description: The timestamp when the client was created.
          example: "2024-03-15T11:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: The timestamp when the client was last updated.
          example: "2024-03-15T15:45:00Z"
      required:
        - name
    # Define the Technician Schema
    Technician:
      type: object
      properties:
        id:
          type: integer
          description: The technician ID.
          example: 1
        name:
          type: string
          description: The name of the technician.
          example: "John Doe"
        role:
          type: string
          description: The role of the technician.
          example: "Electrician"
        availability:
          type: string
          description: The availability status of the technician.
          example: "Available"
        skills:
          type: array
          description: An array of skills the technician possesses.
          items:
            type: string
          example:
            - "Electrical Wiring"
            - "Troubleshooting"
        contact_details:
          type: object
          description: Contact information for the technician.
          properties:
            email:
              type: string
              format: email
              example: "john.doe@example.com"
            phone:
              type: string
              example: "555-987-6543"
        created_at:
          type: string
          format: date-time
          description: The timestamp when the technician was created.
          example: "2024-03-15T12:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: The timestamp when the technician was last updated.
          example: "2024-03-15T16:30:00Z"
      required:
        - name
    # Define the Equipment Schema
    Equipment:
      type: object
      properties:
        id:
          type: integer
          description: The equipment ID.
          example: 1
        name:
          type: string
          description: The name of the equipment.
          example: "Laptop"
        manufacturer:
          type: string
          description: The manufacturer of the equipment.
          example: "Dell"
        model:
          type: string
          description: The model of the equipment.
          example: "XPS 13"
        serial_number:
          type: string
          description: The serial number of the equipment.
          example: "ABCDEFG1234567"
        condition:
          type: string
          description: The condition of the equipment (e.g., New, Good, Fair, Poor).
          example: "Good"
        quantity:
          type: integer
          description: The quantity of the equipment.
          example: 1
        assigned_to_project:
          type: integer
          description: The ID of the project to which the equipment is assigned (if any).
          example: 5
        created_at:
          type: string
          format: date-time
          description: The timestamp when the equipment was created.
          example: "2024-03-15T13:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: The timestamp when the equipment was last updated.
          example: "2024-03-15T17:15:00Z"
      required:
        - name
    # Define the Tool schema
    Tool:
      type: object
      properties:
        id:
          type: integer
          description: The tool ID.
          example: 1
        name:
          type: string
          description: The name of the tool.
          example: "Hammer"
        type:
          type: string
          description: The type of the tool.
          example: "Hand Tool"
        condition:
          type: string
          description: The condition of the tool (e.g., New, Good, Fair, Poor).
          example: "Good"
        purchase_date:
          type: string
          format: date-time
          description: The date and time the tool was purchased.
          example: "2023-01-10T14:30:00Z"
        assigned_to_project:
          type: integer
          description: The ID of the project to which the tool is assigned (if any).
          example: 3
        assigned_to_technician:
          type: integer
          description: The ID of the technician to which the tool is assigned (if any).
          example: 2
        created_at:
          type: string
          format: date-time
          description: The timestamp when the tool was created in the system.
          example: "2023-01-10T14:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: The timestamp when the tool was last updated.
          example: "2023-01-15T09:45:00Z"
      required:
        - name
        - type
    # Define the Notification schema
    Notification:
      type: object
      properties:
        id:
          type: integer
          description: The notification ID.
          example: 1
        user_id:
          type: integer
          description: The ID of the user associated with the notification.
          example: 3
        message:
          type: string
          description: The notification message.
          example: "Task 'Design Database Schema' has been completed."
        read:
          type: boolean
          description: Whether the notification has been read or not.
          example: false
        created_at:
          type: string
          format: date-time
          description: The timestamp when the notification was created.
          example: "2024-03-15T16:00:00Z"
      required:
        - user_id
        - message
    # Define the Settings schema
...
    LandingContent:
      type: object
      properties:
        id:
          type: integer
          description: The landing content ID.
          example: 1
        section_name:
          type: string
          description: The name of the section (e.g., hero, features, about).
          example: "hero"
        content:
          type: object
          description: The content of the section, can be any valid JSON structure.
          example:
            title: "Welcome to Projectiva"
            subtitle: "Your all-in-one project management solution."
            buttonText: "Get Started"
            buttonLink: "/signup"
        updated_at:
          type: string
          format: date-time
          description: The timestamp when the landing content was last updated.
          example: "2024-03-15T19:00:00Z"
      required:
        - section_name
        - content
    # Define the User schema
...